% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allerlei-package.R
\docType{package}
\name{allerlei-package}
\alias{allerlei}
\alias{allerlei-package}
\title{allerlei: Personal Collection of Utility Functions}
\description{
A personal collection of helper functions for data I/O (local and AWS S3), financial and insurance metrics, agricultural and geospatial analysis, data visualisation, and general utilities.
}
\author{
\strong{Maintainer}: Beat Krauer \email{beat.krauer@gmail.com}

}
\keyword{internal}
