% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_collection.R
\docType{data}
\name{cp_col}
\alias{cp_col}
\title{Default colour palette used by plotting functions}
\format{
An object of class \code{character} of length 7.
}
\usage{
cp_col
}
\description{
Default colour palette used by plotting functions
}
\keyword{datasets}
